Support undo/redo in house design mode
