+++ src/elona/building.hpp
@@ -9,7 +9,6 @@ void initialize_home_adata();
 TurnResult build_new_building();
 TurnResult show_house_board();
 void addbuilding(int related_town_quest_id, int building_type, int x, int y);
-void start_home_map_mode();
 void show_home_value();
 void show_shop_log();
 void try_extend_shop();
