+++ src/elona/mapeditor/design_mode.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+
+
+namespace elona
+{
+namespace mapeditor
+{
+
+void start_home_map_mode();
+
+}
+} // namespace elona
